hexes done
