+++ Mother Earth/Assets/Scripts/Controllers/BasicController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10949f7f4c977fa43845661254e212d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
