+++ Mother Earth/Assets/Scripts/Controllers/EnvController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00a9e6d757fca8544a8933edc90a0fef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
